requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
streamlit==1.35.0
pydantic==1.10.14
numpy==1.26.4
scipy==1.11.4
librosa==0.10.2.post1
soundfile==0.12.1
jiwer==3.0.1
tqdm==4.66.4
transformers==4.41.2
onnxruntime==1.17.3
onnxruntime-gpu==1.17.3
torch==2.4.0
tensorflow==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
TensorFlowTTS==0.15.4

PyYAML==6.0.1
